requirements.txt
alembic==1.14.0
backoff==2.2.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Migrate==4.0.7
Flask-Simple-GeoIP==0.2.4
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
ipaddress==1.0.23
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
mysqlclient==2.2.6
numpy==2.1.3
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.30.0
simple-geoip==0.1.1
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==0.18.0
urllib3==2.2.3
user-agents==2.2.0
Werkzeug==3.1.2
WTForms==3.1.2
WTForms-SQLAlchemy==0.4.2
